requirements.txt
# PyTorch with CUDA 11.8 support (RTX 3060)
torch==2.1.0+cu118
torchvision==0.16.0+cu118

# Deep Learning
timm==0.9.12

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Visualization
matplotlib==3.8.0
seaborn==0.13.0

# Metrics & Evaluation
scikit-learn==1.3.2

# Progress bars
tqdm==4.66.1

# Jupyter
notebook==7.0.6
ipykernel==6.26.0

# Utils
pyyaml==6.0.1